Added New PreResult Function
Refactored the preresult function to take in each row and then a second function for all rows.

Allows more flexibility for adjusting values and design space.

Created a parent class for the CrimRdsEvaluator that all the Evaluators inherit from incase I need to add functionality that's the same across each of them.